+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - teste de funcionamento da função MAP
 - teste de funcionamento da função filter
 - teste de funcionamento da função reduce
+- exercício erros
 
 ## 1.1.0 / 15/06/2022
 - resolução do desafio de comparação usando operadores e sintaxe